pyproject.toml
[project]
name = "pylac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.9"
dependencies = [
    "pygame-ce>=2.5.2",
]

[project.scripts]
pylac-pygame = "pylac:main"
